Update API neural base
